Airdrop/serializers.py
# airdrop/serializers.py

from rest_framework import serializers
from .models import Participant

class ParticipantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Participant
        exclude = ('pin',) 
